requirements/back-end.txt
Django
djangorestframework
django-filter
drf-yasg
markdown
requests
python-dotenv
django-cors-headers
django-extensions
Werkzeug
pyOpenSSL